This is an experiment trying to replace `updateNode(node)` with `updateNodeAtPath(path, node)`
It shows that the `memerize` of `getAncestors(key)` is so efficient that `updateNodeAtPath`, surprisingly, slow down the slateJS.

Benchmark
```
$ mkdir -p ./tmp && babel-node ./node_modules/.bin/_matcha --reporter ./support/benchmark-reporter ./packages/*/benchmark/index.js > ./tmp/benchmark-comparison.json && babel-node ./support/benchmark-compare

  benchmarks
    html-serializer
      deserialize
        9.67 → 9.56 ops/sec
      serialize
        1297.79 → 1278.01 ops/sec
    plain-serializer
      deserialize
        628.79 → 664.33 ops/sec
      serialize
        10893.18 → 9974.71 ops/sec
    rendering
      normal
        131.94 → 124.00 ops/sec
    changes
      delete-backward
        162376.05 → 142525.50 ops/sec
      delete-forward
        30069.68 → 32136.36 ops/sec
      insert-text-by-key-multiple
        110.94 → 105.38 ops/sec
      insert-text-by-key
        1448.67 → 1383.92 ops/sec
      insert-text
        1250.24 → 1278.05 ops/sec
      normalize
        2302.14 → 2272.59 ops/sec
      split-block
        31.68 → 30.26 ops/sec
    models
      from-json
        136.29 → 136.50 ops/sec
      get-blocks-at-range
        1186642.28 → 947325.76 ops/sec
      get-blocks
        4714482.86 → 3964856.56 ops/sec
      get-characters-at-range
        979130.16 → 989907.79 ops/sec
      get-characters
        3557471.76 → 3595061.79 ops/sec
      get-inlines-at-range
        1028974.03 → 1013665.48 ops/sec
      get-inlines
        3613919.73 → 3501082.30 ops/sec
      get-leaves
        2798395.10 → 2831306.75 ops/sec
      get-marks-at-range
        1010496.40 → 1016599.82 ops/sec
      get-marks
        3528662.99 → 3402724.27 ops/sec
      get-path
        1305148.78 → 1317090.30 ops/sec
      get-texts-at-range
        1150373.86 → 1243870.35 ops/sec
      get-texts
        3888650.71 → 4626783.06 ops/sec
      has-node-multiple
        176895.60 → 196075.94 ops/sec
      has-node
        1340035.87 → 1378576.69 ops/sec
      to-json
        3614.20 → 3624.90 ops/sec
      update-node
        16461.21 → 15012.86 ops/sec
```